fix(kubewarden): Kubewarden moved the compose file.
Recently the Kubewarden project moved the compose files under the
updatecli directory to keep all of them at the same place. However, this
policy has the file path hard coded. Therefore, this commit adds a new
value to allow defining the file path of the compose file where the
policy should be updated.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>